Move formatting code to trait
